neutralts::bif
Module parse_bif_param

{:param; … :}

Set custom parameters, no output in set, output value in get.

{:param; name >> value :} {:* set *:}
{:param; name :}          {:* get *:}

Modifiers:

{:^param; ... :}

No flags

Usage

The parameters must necessarily be defined within and at the beginning of the bif “code”.

{:code;
    {:param; name1 >> value :}
    {:param; name2 >> value :}
    ...
:}

The parameters have block scope, will be inherited and recover their value:

{:code;
    {:param; name >> 1 :}
    {:code;
        {:param; name >> 2 :}
    :}
    {:* "name" will recover their value *:}
:}

Example of passing arguments to template:

{:code;
    {:param; option >> value :}
    {:include; foo.tpl :}
:}

Example of passing arguments to snippet:

{:code;
    {:param; option >> blue :}
    {:snippet; foo :}
:}

The snippet may be:

{:snippet; foo >>
    The option is: {:param; option :}
:}