neutralts

Module constants

Structs

NEUTRAL_JS

JavaScript for {:fetch; … :}

STATUS_CODES

HTTP status codes

Constants

BACKSPACE

html entity for Backspace

BIF_ALIAS_LIST

bif alias list because some bifs have no name

BIF_ALLOWED

Identify bif allow

BIF_ARRAY

Bif delimiters

BIF_CLOSE

Bif delimiters

BIF_CLOSE0

BIF_CLOSE 0 as bytes

BIF_CLOSE1

BIF_CLOSE 1 as bytes

BIF_CLOSE_B

BIF_CLOSE as bytes

BIF_CODE

Bif delimiters

BIF_CODE_B

BIF_CODE as bytes

BIF_COMMENT

Bif comment

BIF_COMMENT_B

BIF_COMMENT as bytes

BIF_COMMENT_CLOSE

Bif comment close

BIF_COMMENT_CLOSE_B

BIF_COMMENT_CLOSE as bytes

BIF_COMMENT_OPEN

Bif comment open

BIF_COMMENT_OPEN_B

BIF_COMMENT_OPEN as bytes

BIF_DELIM

Bif delimiters

BIF_DELIM_B

delim as bytes

BIF_LIST

bif list

BIF_MOD_FILTER

modifier filter

BIF_MOD_NEGATE

modifier not

BIF_MOD_SCOPE

modifier scope

BIF_MOD_UPLINE

modifier upline

BIF_NAME

Bif delimiters

BIF_OPEN

Bif delimiters

BIF_OPEN0

BIF_OPEN 0 as bytes

BIF_OPEN1

BIF_OPEN 1 as bytes

BIF_OPEN_B

BIF_OPEN as bytes

BIF_SANITIZE_CLOSE

replacement for BIF_CLOSE sanitation

BIF_SANITIZE_OPEN

replacement for BIF_OPEN sanitation

BIF_VAR

Identify bif var as it has no name

CRLF

html entity for crlf

DEFAULT_OBJ_CALLBACK

DEFAULT_OBJ_ENGINE

DIV_FETCH_AUTO

HTML container auto for {:fetch; … :}

DIV_FETCH_CLICK

HTML container click for {:fetch; … :}

DIV_FETCH_FORM

HTML container form for {:fetch; … :}

DIV_FETCH_NONE

HTML container none for {:fetch; … :}

DIV_FETCH_VISIBLE

HTML container visible for {:fetch; … :}

EMPTY

empty

EMPTY_STRING

Empty string could be different in different environments and could be an HTM entity. It is also used to make it clear that it is an empty string.

FALSE

false

NULL

html entity for null

REDIR_JS_REDIRECT_SELF

JavaScript script that redirects to a new URL in the self page, used in {:redirect; The placeholder ‘{}’ should be replaced with the destination url.

REDIR_JS_REDIRECT_TOP

JavaScript script that redirects to a new URL in the top page, used in {:redirect; The placeholder ‘{}’ should be replaced with the destination url.

REDIR_JS_RELOAD_SELF

JavaScript script reloads the current, used in {:redirect;

REDIR_JS_RELOAD_TOP

JavaScript script reloads the current top page, used in {:redirect;

SNIPPETS_FILES

To detect the template files that may contain snippet

SPACE

html entity for space

TRUE

true

UNPRINTABLE

html entity for {:;:}

VERSION

neutralts version